- 调整代码

+++ Engine/Renderer/Vulkan/src/VKTexture2D.cpp
@@ -82,9 +82,9 @@ bool CVKTexture2D::TransferTexture2D(GfxPixelFormat pixelFormat, int level, int 
 	m_transferRegions[level].bufferOffset = 0;
 	m_transferRegions[level].bufferRowLength = 0;
 	m_transferRegions[level].bufferImageHeight = 0;
-	m_transferRegions[level].imageSubresource = { VK_IMAGE_ASPECT_COLOR_BIT, level, 0, 1 };
+	m_transferRegions[level].imageSubresource = { VK_IMAGE_ASPECT_COLOR_BIT, (uint32_t)level, 0, 1 };
 	m_transferRegions[level].imageOffset = { xoffset, yoffset, 0 };
-	m_transferRegions[level].imageExtent = { width, height, 1 };
+	m_transferRegions[level].imageExtent = { (uint32_t)width, (uint32_t)height, 1 };
 	m_transferBuffer[level].assign((uint8_t *)data, (uint8_t *)data + size);
 
 	return true;
@@ -111,9 +111,9 @@ bool CVKTexture2D::TransferTexture2DCompressed(GfxPixelFormat pixelFormat, int l
 	m_transferRegions[level].bufferOffset = 0;
 	m_transferRegions[level].bufferRowLength = 0;
 	m_transferRegions[level].bufferImageHeight = 0;
-	m_transferRegions[level].imageSubresource = { VK_IMAGE_ASPECT_COLOR_BIT, level, 0, 1 };
+	m_transferRegions[level].imageSubresource = { VK_IMAGE_ASPECT_COLOR_BIT, (uint32_t)level, 0, 1 };
 	m_transferRegions[level].imageOffset = { xoffset, yoffset, 0 };
-	m_transferRegions[level].imageExtent = { width, height, 1 };
+	m_transferRegions[level].imageExtent = { (uint32_t)width, (uint32_t)height, 1 };
 	m_transferBuffer[level].assign((uint8_t *)data, (uint8_t *)data + size);
 
 	return true;
